requirements.txt
streamlit==1.28.0
openai==0.28.1
pandas==2.1.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3
torch==2.1.0
transformers==4.35.2
tf-keras==2.15.0
